[Status Panel] Added methods to update the status labels.
The 'updateStatus' and 'updateNumberOfImages' methods allow to interact
with the status panel to update the displayed content.

Added calls to those methods in the main window class.

Minor fix to the timeline in the serial simulator (the number of images
now only increases after the SOE signal has been triggered).